+++ src/libultra/io/sirawread.c
@@ -4,6 +4,6 @@ s32 __osSiRawReadIo(void* devAddr, u32* dst) {
     if (__osSiDeviceBusy()) {
         return -1;
     }
-    *dst = HW_REG((uintptr_t)devAddr, uintptr_t);
+    *dst = HW_REG((uintptr_t)devAddr, u32);
     return 0;
 }
